Initial import of source files
ZonesWindow is now working with the java-gnome 4.0 'treeview' branch,
but all of the code relating to Accelerators and the popup Menu has
been commented out as we don't have effective coverage for that yet.

MeetingWindow has been partially converted to 4.0, but is not yet
hooked up and many widgets are missing as yet.