Improve logic in TimeZoneHelper
Fix several logic bugs in the code related to detecting the
/etc/localtime symlink. Also, ignore case when considering the OS name
when retrieved from the os.name property.

Augment documentation describing the sequence followed in
getUserTimeZone().

Change fallback return value to a "" empty string; no need to throw an
Exception in Place's static initializer, really, and we might as well
just get on with setting that default fallback.