Improve timer algorithm
Rather than waking once per second and testing for time % 60 == 0,
instead work out the number of miliseconds to the next minute (ie,
60000 - time % 60000) and sleep for that long. MUCH better.

Unfortunately, this means that on a suspend / resume cycle, the timer
will be out by up to 1-59 seconds (30 +/- 29) and will display the
wrong value until the clock Thread returns from the sleep() call.

Bump to Java 1.6 to see if is still misbehaves this way. It does.  Not
waking up prematurely (at least with InterruptedException) seems like
a Java bug to me.