Switch from cafesociety to shell
Change top level configure to output a single line shell script
instead of a cafesociety keyfile. Nothing fancy, just a straight
forward exec statement and the java command line. Shell is slightly
more widely available, and we'll keep cafesociety for the linking
project it was meant to be.

Improve ignore list, and version bump to 0.5.3