Introduce the idea of a home location being different from current location
When travelling it seems that most of the time what I needed slashtime
for was to tell what time it was "at home", ie, where I came from.
Hunting for that city in the list when my eye was so used to seeking
the icon marking the location was hard.

So, change isLocal() to isHome() and add a new isLocal(), along with
supporting infrastructure. Hard wire the home location while testing.

Use a former GNOME "start here" icon as the image for the current
location.