Fix sort ordering to be stable
Wwe have to go to some additional trouble to ensure a stable sort
ordering. Originally we just used halves, but we were having a problem
with TreeView repainting and worse resorting but back to the original
state. Ugly.

The root cause seems to have been that several rows had the same
sort number (halvesSinceMidnight). To give each Place a unique sort
descriminator, we add an incrementing counter to, after first scaling
that number up by 100 so zone ordering by halves is perserved.