UI improvements
When MeetingWindow is active, return ordering to sorting by offset and
reset minutes to :00. Change window border to red as a visual cue that
the ZonesWindow is not displaying current time information.

Fix focus issues, using LEAVE_NOTIFY_EVENT instead of FOCUS_OUT_EVENT
and finessing matters so that when the popup window is activated the
selection is maintained.