Allow user to specify home location on command line
If offered, the first (and only) option on the command line will be
processed as being the zone name of the user's home (as opposed to
current) location.

The 'home' feature is still a bit of a kludge, but at least it is no
longer hard coded to Australia/Sydney :)

Remove obscelete launcher script.

Reduce unncessary public visibility on Place methods.