Preliminary internationalization of slashtime
Initialize internationalization library in Master.

Mark all UI strings as translatable with _() function.

Build creates a PO template at tmp/i18n/slashtime.pot; not sure if
that's the right place for it, or if it should be under version
control.

Add .po file for fr_CA as an example translation to new top level
directory po/.

Update Makefile to compile .mo files for listed domains. Still needs
to be made dynamic so as to simply react to .po files present in po/.