Refactor constructor into multiple methods
The ZonesWindow class is very long as there is a great deal involved
in constructing and managing a Window that is the application's
primary presentation and also contains a TreeView.


In particular, using outline function in an IDE showed the singal
connection (due to the inner classes) but made navigating and
following the code fairly difficult.

So, instead of having one monster constructor, refactor it into a
series of calls to appropriately named methods. This improves the
readibility and ability to jump around, at the cost of the various
Widgets and other GTK fields not being final.