Refactor into multple layers with package separation
Move the Java sources from a single com.operationaldynamics.slashtime
package to slashtime.{ui,client,services,domain,util}, being one
package per layer following enterprise architecture patterns.

Create new class slashtime.ui.UserInterface to be instantiated and
assigned to Master.ui, moving the GUI construction code that was in
Master to this class.

As ever, we are forced to play some games to make reentry points
visible. Some of this could be short cut, but it is there because it
facilitates scaling up massively.