Restrict the minute slider to quarter hours
As a usability improvement, restrict the values of the minute HScale
to 0, 15, 30, 45  (and, still looking somewhat clumsy, 59) minutes.
This prevents people from having to fiddle around to get to the
cardinal points of the minute clock, which is much better.

This was implemented in a somewhat questionable fashion; there may be
a better (more GTKish) way to do this.